app: Move store into ee
